Remove the ArgFilterDirective interface and related functionality
