+++ docs/master/api-reference/directives.md
@@ -2109,8 +2109,6 @@ input DateRange {
 The type of the input value this is defined upon should be
 an `input` object with two fields.
 
-_Attention: To use this new definition style, set the config `new_between_directives` in `lighthouse.php`._
-
 ### Definition
 
 ```graphql
@@ -2126,21 +2124,6 @@ directive @whereNotBetween(
   """
   key: String
 ) on ARGUMENT_DEFINITION | INPUT_FIELD_DEFINITION
-```
-
-### Old syntax
-
-_Attention: This following use is deprecated and will be removed in v4._
-
-_Note: You will need to add a `key` to the column to want to query for each date_
-
-```graphql
-type Query {
-    posts(
-        createdAfter: Date! @whereBetween(key: "created_at")
-        createdBefore: String! @whereBetween(key: "created_at")
-    ): [Post!]! @all
-}
 ```
 
 ## @whereConstraints
@@ -2249,8 +2232,6 @@ Verify that a column's value lies outside of two values.
 The type of the input value this is defined upon should be
 an `input` object with two fields.
 
-_Attention: To use this new definition style, set the config `new_between_directives` in `lighthouse.php`._
-
 ```graphql
 type Query {
     posts(
@@ -2279,21 +2260,6 @@ directive @whereNotBetween(
   """
   key: String
 ) on ARGUMENT_DEFINITION | INPUT_FIELD_DEFINITION
-```
-
-### Old syntax
-
-_Attention: This following use is deprecated and will be removed in v4._
-
-_Note: You will need to add a `key` to the column to want to query for each date_
-
-```graphql
-type Query {
-    users(
-        bornBefore: Date! @whereNotBetween(key: "created_at")
-        bornAfter: Date! @whereNotBetween(key: "created_at")
-    ): [User!]! @all
-}
 ```
 
 ## @with
