+++ src/Schema/Directives/WhereNotBetweenDirective.php
@@ -2,14 +2,12 @@
 
 namespace Nuwave\Lighthouse\Schema\Directives;
 
-use Nuwave\Lighthouse\Support\Contracts\ArgFilterDirective;
+use Nuwave\Lighthouse\Support\Contracts\ArgBuilderDirective;
 
-/**
- * @deprecated in favour of
- * @see \Nuwave\Lighthouse\Between\WhereNotBetweenDirective
- */
-class WhereNotBetweenDirective implements ArgFilterDirective
+class WhereNotBetweenDirective extends BaseDirective implements ArgBuilderDirective
 {
+    const NAME = 'whereNotBetween';
+
     /**
      * Name of the directive.
      *
@@ -17,30 +15,21 @@ class WhereNotBetweenDirective implements ArgFilterDirective
      */
     public function name(): string
     {
-        return 'whereNotBetween';
+        return self::NAME;
     }
 
     /**
+     * Apply a "WHERE NOT BETWEEN" clause.
+     *
      * @param  \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder  $builder
-     * @param  string  $columnName
-     * @param  mixed  $value
+     * @param  mixed  $values
      * @return \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder
      */
-    public function applyFilter($builder, string $columnName, $value)
+    public function handleBuilder($builder, $values)
     {
-        return $builder->whereNotBetween($columnName, $value);
-    }
-
-    /**
-     * Does this filter combine the values of multiple input arguments into one query?
-     *
-     * This is true for filter directives such as "whereBetween" that expects two
-     * different input values, given as separate arguments.
-     *
-     * @return bool
-     */
-    public function combinesMultipleArguments(): bool
-    {
-        return true;
+        return $builder->whereNotBetween(
+            $this->directiveArgValue('key', $this->definitionNode->name->value),
+            $values
+        );
     }
 }
