+++ src/Schema/Factories/FieldFactory.php
@@ -9,10 +9,8 @@ use GraphQL\Type\Definition\InputType;
 use GraphQL\Type\Definition\ListOfType;
 use Nuwave\Lighthouse\Support\Pipeline;
 use Nuwave\Lighthouse\Execution\Builder;
-use Nuwave\Lighthouse\Schema\AST\ASTHelper;
 use GraphQL\Type\Definition\InputObjectType;
 use Nuwave\Lighthouse\Execution\ErrorBuffer;
-use Nuwave\Lighthouse\Execution\QueryFilter;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use GraphQL\Language\AST\InputValueDefinitionNode;
 use Nuwave\Lighthouse\Schema\Values\ArgumentValue;
@@ -23,7 +21,6 @@ use Nuwave\Lighthouse\Schema\Directives\SpreadDirective;
 use Nuwave\Lighthouse\Support\Contracts\HasArgumentPath;
 use Nuwave\Lighthouse\Support\Contracts\ProvidesResolver;
 use Nuwave\Lighthouse\Support\Traits\HasResolverArguments;
-use Nuwave\Lighthouse\Support\Contracts\ArgFilterDirective;
 use Nuwave\Lighthouse\Support\Contracts\ArgBuilderDirective;
 use Nuwave\Lighthouse\Support\Contracts\ArgDirectiveForArray;
 use Nuwave\Lighthouse\Support\Contracts\ArgValidationDirective;
@@ -68,12 +65,6 @@ class FieldFactory
      * @var \Nuwave\Lighthouse\Execution\Builder
      */
     protected $builder;
-
-    /**
-     * @var \Nuwave\Lighthouse\Execution\QueryFilter
-     * @deprecated
-     */
-    protected $queryFilter;
 
     /**
      * @var array
@@ -172,8 +163,6 @@ class FieldFactory
                 $this->validationErrorBuffer = (new ErrorBuffer)->setErrorType('validation');
                 $this->builder = new Builder;
 
-                $this->queryFilter = QueryFilter::getInstance($this->fieldValue);
-
                 $argumentValues->each(
                     function (ArgumentValue $argumentValue): void {
                         $this->handleArgDirectivesRecursively(
@@ -210,10 +199,6 @@ class FieldFactory
                         );
                     }
                 }
-
-                $this->builder->setQueryFilter(
-                    $this->queryFilter
-                );
 
                 // The final resolver can access the builder through the ResolveInfo
                 $this->resolveInfo->builder = $this->builder;
@@ -388,18 +373,6 @@ class FieldFactory
             if ($directive instanceof ArgBuilderDirective) {
                 $this->builder->addBuilderDirective(
                     $astNode->name->value,
-                    $directive
-                );
-            }
-
-            if ($directive instanceof ArgFilterDirective) {
-                $argumentName = $astNode->name->value;
-                $directiveDefinition = ASTHelper::directiveDefinition($astNode, $directive->name());
-                $columnName = ASTHelper::directiveArgValue($directiveDefinition, 'key', $argumentName);
-
-                $this->queryFilter->addArgumentFilter(
-                    $argumentName,
-                    $columnName,
                     $directive
                 );
             }
