+++ tests/Integration/Execution/BuilderTest.php
@@ -7,16 +7,6 @@ use Tests\Utils\Models\User;
 
 class BuilderTest extends DBTestCase
 {
-    protected function getEnvironmentSetUp($app)
-    {
-        parent::getEnvironmentSetUp($app);
-
-        /*
-         * @deprecated once the new directives become the standard
-         */
-        $app['config']->set('new_between_directives', true);
-    }
-
     /**
      * @var \Illuminate\Database\Eloquent\Collection<\Tests\Utils\Models\User>
      */
